requirements.txt
# Core Agent Framework
google-adk==1.14.0
langchain==0.3.27
langchain-community==0.3.27
langgraph>=0.2.70

# Google Cloud & AI
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.5.0

# Cloud Firestore 
google-cloud-firestore>=2.16.0
google-cloud-datastore>=2.19.0

# FastAPI Backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# MCP Tools
mcp>=1.0.0

# Currency & Financial APIs
requests>=2.31.0
httpx>=0.27.0

# Multilingual NLP
langdetect>=1.0.9

# Environment & Utilities
python-dotenv>=1.0.0